moved options under /component/options
