+++ app/screens/post_options/components/options/copy_link_option.tsx
