+++ app/screens/post_options/components/options/copy_text_option.tsx
