+++ app/screens/post_options/components/options/follow_option.tsx
